refactor: consolidate E2E tests — replace stubs with full suite
Remove the old tests/e2e/ directory which contained only stub tests
(no assertions, no navigation, no UI interaction) and rename
tests/e2e_full/ → tests/e2e/ as the single E2E test suite.

The old tests/e2e/ had:
- test_basic_navigation: just ran 20 frames
- test_wallet_balance_rendering: just ran 30 frames
- test_wallet_state_initialization: just ran 20 frames
- TestHarness struct: never used

All of this is superseded by the comprehensive phase-based E2E suite
and the kittest interaction tests.
