--- tests/e2e/helpers.rs
@@ -1,36 +1,0 @@
-//! E2E Test Helpers
-//!
-//! This module provides shared utilities for E2E testing, including:
-//! - Test harness setup
-//! - Common test fixtures
-
-/// Create a minimal test harness for E2E tests
-#[allow(dead_code)]
-pub struct TestHarness {
-    pub runtime: tokio::runtime::Runtime,
-}
-
-impl TestHarness {
-    pub fn new() -> Self {
-        let runtime = tokio::runtime::Runtime::new().expect("Failed to create tokio runtime");
-        Self { runtime }
-    }
-}
-
-impl Default for TestHarness {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_harness_creation() {
-        let harness = TestHarness::new();
-        // Just verify we can create the harness without panicking
-        drop(harness);
-    }
-}
