--- tests/e2e/mod.rs
@@ -1,9 +1,0 @@
-//! E2E Test Suite for Dash Evo Tool
-//!
-//! This module contains end-to-end tests that verify complete user journeys.
-//! The tests use egui_kittest to simulate the UI and verify that screens
-//! render and behave correctly.
-
-mod helpers;
-mod navigation;
-mod wallet_flows;
