+++ tests/e2e/phases/phase_02_wallet.rs
