+++ tests/e2e/phases/phase_03_platform.rs
