+++ tests/e2e/phases/phase_04_tokens.rs
