+++ tests/e2e/phases/phase_smoke.rs
